pyproject.toml
[project]
name = "ctrl-alt-de-leash"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "py-cord>=2.6.1",
    "python-dotenv>=1.1.0",
]
